backend/requirements.txt
Flask==3.0.2
numpy==1.26.4
scipy==1.12.0
qiskit==1.0.2
qiskit-optimization==0.6.1
python-dotenv==1.0.1